refactor: adiciona uma classe util para uso com specification
